Add a LithoHotswapManager
Summary: This is the first part in enabling hotswap to occur in Litho (the next part will provide the annotation processor changes to enable this). Users will need to set up a ClassLoader that they want to be used to load the spec class, and to call `onHotswap` when a hotswap has taken place.

Reviewed By: marco-cova

Differential Revision: D13535166

fbshipit-source-id: 227debf85176d72e5ba31e0dcf3c64959bab32ff