pass the parent component context while resolving nest tree
Summary:
The implementation of `LayoutState.resolveNestedTree` calls `LayoutState#createAndMeasureTreeForComponent` with the context of the component it is resolving whereas it expects the context to belong to the parent because the global key is generated with respect to the parent of the component.

Currently, this messes up the global key generation and hence the state updates for the root component of the nested tree. An example for the below nested tree; where `B` is a nested tree:

```
A
|-- B
     |--C
```

Expected global keys:

* `A`
* `A,B`
* `A,B,C`

Generated global keys:

* `A`
* `A,B,B`
* `A,B,B,C`

Even though we will disallow rewriting the global key, which would solve this specific problem; but passing the incorrect context may cause other side effect; now or in the future.

Reviewed By: marco-cova

Differential Revision: D13320870

fbshipit-source-id: 9532e387b939354c08567de14463bfb7172da2b5