use a shallow copy for yoga measure pass
Summary:
Use a shallow copy of the root component for the Yoga measure pass. This will ensure that component is resolved without any side effects caused by methods previously called on it.

In this case the global key is reset so that a new global key is not generated for the the root component; which would also change the global keys which will be set to it descendants. This would break state updates.

Reviewed By: astreet

Differential Revision: D13326753

fbshipit-source-id: 571e856c94a09c1783e9b238ea91ad28f38c508e