Fix hotswap to recalculate every layout
Summary: Right now, we force a layout on every LithoView when we hotswap. However, there can be ComponentTrees (e.g. in a Recycler) that have had a layout calculated but are not attached to a LithoView. We need to make sure that these get the updated code too. The way to do that is to force a layout on every ComponentTree, either via the LithoView that it is attached to, or by setting a flag directly on the ComponentTree to make it relayout itself when it gets attached to the LithoView.

Reviewed By: astreet

Differential Revision: D13554865

fbshipit-source-id: 092b42353eef24660e340ea0b0681ea3e86ba660