Made clickable spans default to a 'Button' role for accessibilty
Summary:
Clickable Spans in Litho's Text component can be made accessible by setting the accessibleClickableSpans prop to 'true', but this only made them focusable and did not allow you to customize their role.

You could customize their announcement by using the AccessibleClickableSpan class rather than the ClickableSpan class, and setting an accessibilityDescription for the span, but this forces people to add text usually associated with roles into the content description, which is a bad pattern.

This adds the ability to set a role on AccessibleClickableSpan's as well as defaults the role for ones not set to be "Button", which is the common case for clickable text within a paragraph.

Reviewed By: IanChilds

Differential Revision: D13053454

fbshipit-source-id: 6e6af29b5e48e9208ee167e8c41470b480a02745