Reset pressed state for ComponentHost when it's being detached
Summary:
There is a certain scenario where `ComponentHost` might retain previous drawable state before being released to the pool, and next time it is reused the wrong state is applied to its drawables.

Consider this case:
Let's say we have `PillComponent` with this layout:
```
ContainerA (has click listener)
|
|________ Button
|         (has statelistdrawable as background which draws different color on pressed state,
|          wrapped in view,
|          also duplicates parent's state)
|
|________ Icon
```

And multiple such components are added as children to `VerticalScroll`. Let's say we have a logic to remove items from `VerticalScroll` as we click them. Then, when removing last item from `VerticalScroll` `Button` component's mountable content (`ComponentHost`) will retain pressed state and will be released to pool with pressed state. Next time `Button` component is reused, it wrongly applies pressed state. Note that this behavior happens when user taps the item quickly (if view is "prepressed" (http://androidxref.com/7.0.0_r1/xref/frameworks/base/core/java/android/view/View.java#11102)):
Here is the sequence:
1) User taps quickly on `Button`. `ComponentHost` corresponding to `Button` gets `ACTION_DOWN` and as it doesn't handle click (its `onTouchEvent` returns `false`), touch event goes to its parent - `ComponentHost` corresponding to `ContainerA`, which registers `CheckForTap` runnable. For now both `ComponentHost`s (CH) "pressed" state are false, but both are in "prepressed" state.
2) `ACTION_UP` for `ContainerA`'s CH is triggered before `CheckForTap` is triggered (`ACTION_UP` came less than 100ms after `ACTION_DOWN`). It calls `setPressed(true)` (http://androidxref.com/7.0.0_r1/xref/frameworks/base/core/java/android/view/View.java#11116), which calls `refreshDrawableState` and pressed state is propagated to children who duplicates parent's state. As `Button`'s CH does that it receives pressed state, so the `Button` is drawn with pressed state. The CH that received ACTION_UP (`ContainerA`) registers pending callback to remove pressed state (`UnsetPressedState` (http://androidxref.com/7.0.0_r1/xref/frameworks/base/core/java/android/view/View.java#11142)). However, `Button`'s CH didn't register `UnsetPressedState` as it is not recipient of touch, even though it is in pressed state by now.
3) In `ClickEvent` of the `ContainerA` we do state update to set new list to `VerticalScroll` that excludes clicked item. We call `unmountView` on `Button`'s CH (which still has pressed state). `Button`'s CH's `dispatchDetachedFromWindow` is called, but it cannot clear pressed state as it has no `UnsetPressedState` registered: http://androidxref.com/7.0.0_r1/xref/frameworks/base/core/java/android/view/View.java#15252. That CH goes into pool with pressed state. `ContainerA`'s CH's pressed state is correctly cleared during `unmountView` as on its `dispatchDetachedFromWindow` clears pressed state as it has  `UnsetPressedState` registered.
4) Next time `Button`'s CH is reused it will still contain pressed state, and its `StateListDrawable` picks pressed state content.

Note that if you remove non-last item, this issue doesn't occur because the pressed item is not released to the pool, but reused for next item during mountdiffing. And the last item will be released to the pool which is not the one pressed, therefore it enters pool in clean state.

Reviewed By: astreet

Differential Revision: D13675335

fbshipit-source-id: de116ace90c1443ef96295ef32d7cf5326a1feea