Made clickable spans default to a 'Button' role for accessibilty... take two
Summary:
Clickable Spans in Litho's Text component can be made accessible by setting the accessibleClickableSpans prop to 'true', but this only made them focusable and did not allow you to customize their role.

You could customize their announcement by using the AccessibleClickableSpan class rather than the ClickableSpan class, and setting an accessibilityDescription for the span, but this forces people to add text usually associated with roles into the content description, which is a bad pattern.

This adds the ability to set a role on AccessibleClickableSpan's as well as defaults the role for ones not set to be "Button", which is the common case for clickable text within a paragraph.

**Note: This is a re-do of D13053454, which was reverted due to failing builds.  The only change was switching the buck file to use provided_deps rather than deps.**

Reviewed By: RomanTruba

Differential Revision: D13642644

fbshipit-source-id: e40a13b18c10f7b44d65c4aade21045ef9333b77