Add custom role descriptions to Litho
Summary:
Accessibility announcements on Android generally include three pieces of information; State, ContentDescription/text, Role.  ContentDescription or text has always been customizable, but State and Role were both only available to be customized by setting predefined states or roles that Android's accessibility services would consume and map to their own strings.  This diff allows the role portion of the announcement to be customized.

It should be noted that this new property is *not* a replacement for accessibilityRole.  AccessibilityRole controls both the default string that accessibility services will announce, as well as defines the behavior of that element.  If you build a custom button type that you want to still *work* like a button, but you want it to be announced as "Cool New Button", you should still set an accessibilityRole of "AccessibilityRole.BUTTON", but you should also set an accessibilityRoleDescription of "Cool New Button".

If you don't want any predefined behavior, but do still want a role to be announced for users, you can simply set the accessibilityRoleDescription alone.

Reviewed By: IanChilds

Differential Revision: D8066609

fbshipit-source-id: 460ab3eebb25f4636a2af425e3c9da0d1103e15f