Suppress requestFocus calls while mounting
Summary: If we have a Component that has focus and it's being unmounted, its host.removeView() calls will clear its focus and ask the rootView to give focus to some other view in the hierarchy. However, this creates problem while we are mounting if other Components gets the focus event and we didn't finish to mount/unmount them.

Reviewed By: muraziz

Differential Revision: D13750235

fbshipit-source-id: 5862001347e21396b3f91e90e6f3abca1aacb4e6