Fix bug where root update is lost when pending async layout is incompatible with new measure specs
Summary:
This is a very special and specific bug. It came up in composer in the Pages Manager app. The bug has to do with some logic we added to prevent a sync layout in measure when we're already computing a new compatible layout in the background (e.g., you setRootAsync and then get a measure: you don't want to synchronously recompute a layout with the root you set in setRootAsync). Unfortunately, this didn't handle the following case.

LithoView is in a LinearLayout with weights that causes it to be measured twice (once with UNSPECIFIED and once with AT_MOST):
0. (unmeasured)
1. setRoot(A)
2. measure height UNSPECIFIED -> main thread layout is calculated with height UNSPECIFIED, height is 250
3. measure height AT_MOST 1000 -> main thread layout is still compatible since it was height 250

// User interaction causes an update

4. setRootAsync(B) -> current height spec is AT_MOST 1000, so calculate a bg layoutstate with height spec AT_MOST 1000, height is determined to be 250. bg layout state is updated from the layout thread.
5. same frame (so bg layoutstate isn't applied), measure UNSPECIFIED which is NOT compatible with the bg layout state (because it was measured with AT_MOST 1000). However, the existing main thread layout state is still compatible since it was originally measured for an UNSPECIFIED spec in (2) above. bg layout is discarded, meaning we now don't have an layout pending for the new root.
6. measure with AT_MOST 1000 again, (old) main thread layout is still compatible
7. Old root is still displayed, async layout is lost

The approach here is to just make sure we only skip the sync layout in measure if the async layout is actually compatible with the new specs. Otherwise we will trigger a sync layout as before.

Reviewed By: pasqualeanatriello

Differential Revision: D13782083

fbshipit-source-id: 3b694ec33e6913b08f04bf1a357d8b6b499db819