Create DispatchKeyEvent Trigger to TextInputSpec
Summary: This adds support to dispatch an Android KeyEvent to the underlying EditText by setting a trigger on the spec. This follows a similar pattern to how we can setSelection or setText on the TextInputSpec. This allows external code to communicate KeyEvents such as KeyEvent.KEYCODE_DEL to the EditText.

Differential Revision: D13871325

fbshipit-source-id: 3260d063a5cc9f4b9b56bf68aba438e2d19b9354