Fix release build issue with d8
Summary:
Took me a while to understand what's going on there, but the reason
why d8 fails for `./gradlew :sample:assembleRelease` is that it pulls
in both `lib/fbjni` and the Flipper dependency `com.facebook.flipper:fbjni` and
doesn't know how to deduplicate. This leads to this error:

```
D8: Program type already present: com.facebook.jni.CppException
```

We can manually exclude the transitive dependency from Flipper as we already
have our own version of it.

Reviewed By: muraziz

Differential Revision: D13897597

fbshipit-source-id: 31760f781bdcea5b55b29d0ad64faab7f3e19f68