resolves issue with code generation with generics
Summary: The diff fixes the issue where state update method uses the same generic multiple times. The diff changes the way the `StateGenerator` builds the state update class. Instead of simply adding every generic type to the class it is first collected in a map (with the key set to the name of the generic) and then set on the state update class.

Reviewed By: muraziz

Differential Revision: D13958193

fbshipit-source-id: a94a4b10330b27507a69321e2032781b0252ea53