using useHeightAsBaseline(true) at all places where YogaAlign.BASELINE is used
Summary:
To make behaviour of baseline consistent with flexbox standard, we are removing setting baseline function (which returned the height of component) on every component. This will also enable us to add new features on top of baseline where users can specify which child to use as a baseline.

And to make sure old behaviour, this diff uses useHeightAsBaseline(true) at all places where Align Items BASELINE was used.

Reviewed By: astreet

Differential Revision: D13838761

fbshipit-source-id: 1496ed4e0594b176fe0d6354af42da137457a186