not setting baseline function and also removed the experiment checks
Summary:
[BREAKING CHANGE]
The default behaviour of baseline has changed. Earlier height of component was used as baseline of component , from now on the baseline of first child of component will be the baseline as per the flexbox standards. This behaviour can be overriden by setting a property useHeightAsBaseline(true) on a component.

To make sure old behaviour does not break, All the immediate children of a component which uses alignItems(YogaAlign.BASELINE) will have to override their baseline using useHeightAsBaseline(true).

Reviewed By: astreet

Differential Revision: D13650597

fbshipit-source-id: 438d24af09e7c3c0a5ae49c840a635b6a0795fdb