Set Section information on RenderInfo's Flipper custom attribute
Summary:
If sections debugging in flipper is enabled, DiffSectionSpecs will send information to the RenderInfo about the parent section that created them. Inside the RecyclerBinder we map the View that renders the RenderInfo's ComponentTree to the information set by the DiffSectionSpec.
In Flipper we'll be able to retrieve the Section information from the RenderInfo's view and reconstruct the section hierarchy.
Depends on D13803302

Reviewed By: astreet

Differential Revision: D13803338

fbshipit-source-id: e72be2dce9277c32d6062d028f1800c6e4064807