Make ChoreographerCompat safe to use from any thread
Summary: ChoreographerCompat previously assumed it would only get instances from the main thread (it directly uses Choreographer.getInstance() which returns the thread instance). In order to make it safe to use from any thread, we now support creating it from any thread by either immediately creating the Choreographer as before, or posting a runnable and setting it from the main thread, falling back to a simple handler in the meantime.

Reviewed By: muraziz

Differential Revision: D14064643

fbshipit-source-id: 96b87493408177a13dd174e0a50bd4f90ffd23ce