Check for detach after dispatching NewLayoutStateListener call
Summary: Not totally sure this is the cause of the crash, but it seems like it could well happen: the listener can call into RecyclerBinder code which updates the RecyclerView which may potentially cause the view to detach. I verified in the unit test that we would get the same NPE if this happened, and that it's fixed with this check.

Reviewed By: pasqualeanatriello

Differential Revision: D14149383

fbshipit-source-id: 0df702bec189910cc68ff3870d0c1c94b4eb654f