Expose get triggers for TextInput charsequence and selection state
Summary: Exposing a `getText` trigger allows clients to treat the `TextInput` as a source-of-truth for the text and selection state without having to duplicate that state.  This allows clients to implement (non-atomic) read-modify-write behavior against the text input and ensures that clients don't accidentally hold onto stale state.  We already expose `setX` analogs, so this seems like a natural extension to the existing functionality.

Differential Revision: D14146294

fbshipit-source-id: 8a4191c53dcca87574979bc7ee37a5f6fb511eec