Config for no init range
Summary:
If the config flag is enabled, the only place that can trigger a sync layout to get the item's size is in measure when the size specs can't be used to measure. Otherwise, all usages of initRange will be replaced with a call to estimate a size for the async range and compute range after. Estimating the range size by scheduling a layout to get its size is always done asynchronously because the async range is also computed async.
Will add tests if this approach makes sense.

I wasn't sure how else to separate the two implementations, hope this isn't too horrible to review.

Depends on D14169171

Reviewed By: astreet

Differential Revision: D14169556

fbshipit-source-id: 6e6d482811f7e24a3ad4abf9b17e3534615916bd