Separate measure and range size
Summary:
Part of the proposal to separate the sync layout and estimating the range size in initRange. This shouldn't change any existing behaviour but instead of working with with a RangeCalculationResult the RecyclerBinder will keep the estimated range size and the Size of the sync layout used for measure separately.
Still keeping RangeCalculationResult around for tests but if this whole stack makes sense I'll get rid of it.

Reviewed By: astreet

Differential Revision: D14168941

fbshipit-source-id: 16b147e426d97af26574f2f44df69060b80bb394