Only add visible Sections in Flipper hierarchy
Summary: when highlighting a Section's children we rely on the view bounds of the first and last child to calculate the parent's bounds. If the last child is a Section that's not yet visible, it will not have any bounds information on it. This only adds visible Sections as children. The next diff will have the logic to update a Section's children as the views change on screen

Reviewed By: passy

Differential Revision: D14300226

fbshipit-source-id: 051a7e4fb8aabec3a78e3cdee718d0cb46a4d6ac