Make NodeInfo an interface
Summary:
This converts `NodeInfo` into an interface and moves the current implementation to a new class `DefaultNodeInfo`. In order to make different implementations inter-op and avoiding hard casts the implementation of `copyInto`, `isEquivalentTo` have been changed.

`copyInto` now accepts another `NodeInfo` instead of `InternalNode` and calls the setters instead of directly setting the field values. In addition the `updateWith` method has been removed as it does logically the same operations as `copyInto` but with `this` and `target` reversed. The usages of `updateWith` has been replaced by the appropriate call to `copyInto`.

The implementation of `isEquivalentTo` method has been moved into a static utility method as it can be reused by different implementations of `NodeInfo`, e.x. `SparseNodeInfo` added in the diff above.

Reviewed By: pasqualeanatriello

Differential Revision: D14322874

fbshipit-source-id: 0f7e759c41dc0d3ce0d94356cc4adff5effe5902