add maximum scroll jump
Summary:
When a fling gesture happens, the default behavior of SNAP_TO_CENTER_CHILD scroll too many cards at a time, we would like to test a slow scroll with a smaller number of cards scrolled at a time.
This diff added the litho support for it, I create a new CustomLinearSnapHelper class which extends LinearSnapHelper, the difference between these two is that in CustomLinearSnapHelper we can set a custom threshold for the fling jump. We will use this in HScrollSpec in next diff.

Reviewed By: andyxhz

Differential Revision: D14435037

fbshipit-source-id: 26532fd705dce9f5cc11a4fe9020b354f8c09895