Remove redundant synchronization + correct check for weakref
Summary:
While trying to add new hooks for instrumentation of scheduled tasks, I've noticed that we are doing un-necessary synchronization in ComponentTree and we're also assuming that the WeakReference is not going to be null the second time we access it.

In this diff I'm removing redundant checks and doing a small refactoring to set the correct behavior.

Reviewed By: marco-cova

Differential Revision: D14513887

fbshipit-source-id: e42127ddc69b0827667c779532b272b52ec5e568