Check if StateHandler is null before getting cachedvalue
Summary:
The ComponentTree could have been released before we try to get/update the cached value on the StateHandler. Releasing
the ComponentTree sets the StateHandler = null, so when we try to get the cached value, we throw an npe. This diff
adds a null check before invoking get/put on the StateHandler

Reviewed By: pasqualeanatriello

Differential Revision: D14578527

fbshipit-source-id: 9279228efce827f803d92c446c77e792dacdad74