expose View's clickable property to Litho
Summary: We do this by (1) creating the appropriate getters/setters, (2) being sure to create a View if clickable is set, (3) using the clickable property to control View#setClickable. A sample use case of this is if you have something with a touchHandler (but not a clickHandler), but want it to be marked as tappable by Android accessibility readers.

Reviewed By: marco-cova

Differential Revision: D14497739

fbshipit-source-id: 2d93e6e1b077b6e7fd76a284dc9b9f801ab058a0