Enabling some of the Litho RecyclerBinder tests.
I have extended the Android classes which were being mocked and removed the mocks.

Tried creating shadows of the Android classes but its not the right solution : As per the Robolectric documentation : http://robolectric.org/extending/

"It is important that shadow methods are implemented on the corresponding shadow of the class in which they were originally defined. Otherwise Robolectric’s lookup mechanism will not find them (even if they have been declared on a shadow subclass.) For example, the method setEnabled() is defined on View. If a setEnabled() method is defined on ShadowViewGroup instead of ShadowView then it will not be found at run time even when setEnabled() is called on an instance of ViewGroup."

In many of the tests here Litho is trying to verify the method of super class is being called on a subsclass object. Shadows doesn't support this.