Update HorizontalScroll example to use Lazy State instead of Atomics
Summary: While normal State updates trigger relayout, it's was a good practice to use Atomics as State values, to updates some State that doesn't require relayout. But it's a hack and Lazy State should have been fixed to be useful for such usecases. Now when it is fixed, samples should reflect better approach..

Reviewed By: mihaelao

Differential Revision: D14645352

fbshipit-source-id: d7bfddd911f07969d597f2111a1a2d00022f603b