Make Accessibility Label Nullable
Summary:
After chatting with Brett Lavalla on D14791043 we determined that on Android it makes sense to have a text span with an accessibility role but *without* an accessibility description.

This diff changes the `AccessibleClickSpan` class to reflect this. In this diff I change all `accessibilityDescription` fields to have a `Nullable` annotation with the exception of the `setAccessibilityDescription` method (I think it's sensible to avoid having the function become a no-op).

Reviewed By: usikder

Differential Revision: D14792206

fbshipit-source-id: 7d2da9c5b83fac54e078b6ec3e234ad2750fc012