fix nested scrolling behavior
Summary:
in ACTION_DOWN, our scroll view needs to tell the parent StoryviewerGestureInterceptingLayout not to intercept the gesture if the user starts moving their finger in a direction the scroll view can scroll. this allows the LithoScrollView to handle the scrolling of the content. on the other hand, if they're at the top of the content, we don't override dragging down, because then we want StoryviewerGestureInterceptingLayout to intercept and handle that gesture so that drag-to-dismiss works.

the reason this needs to happen in onInterceptTouchEvent and not onTouchEvent, is because a child view in the content could decide to handle the onTouchEvent (like a clickable attachment), in which case LithoScrollView won't get the onTouchEvent.

Differential Revision: D14738123

fbshipit-source-id: 66429d57edd97dc6e451de5a6a914bf99a89dc97