Move targetSdk configuration from AndroidManifest to build.gradle. Fix running Unit test from Gradle (#529)
Summary:
Pull Request resolved: https://github.com/facebook/litho/pull/529

Gradle sync is broken, because build config doesn't support setting `minSdkVersion` and `targetSdkVersion` in
AndroidManifest for a long time. We need to move these config params to `build.gradle`, but this will break our
Unit test execution, because of the old Robolectric version which relies on AndroidManifest. Fix TestRunner
accordingly. Closes #515

Reviewed By: passy

Differential Revision: D14853883

fbshipit-source-id: b695521fd8512304443891c4b1745a2f102e6422