Handle null LayoutState returned by LayoutStateFutures
Summary:
Right now LayoutStateFutures can return a null LayoutState if they get released. This can happen especially when turning on cancelation.
This diff still crashes in illegal situations when LayoutState is null, but exists without crash if the LayoutStateFuture was released by cancelation and we don't need to use the LayoutState

Reviewed By: pasqualeanatriello

Differential Revision: D14913174

fbshipit-source-id: 1642942848a2a37062ef4630a5a33225c7a526ae