BREAKING: Release ComponentTree through binder when RecyclerCollectionComponent's @OnDetached is called.
Summary:
When RecyclerCollectionComponent's onDetached() callback is called, we should release ComponentTrees in the sections as well. Add a method called `release` in Binder.java that RecyclerCollectionComponent can access it to release ComponentTrees under the hood.

The release() method in ComponentTreeHolder has been removed in D14183719, but I think we need to add it back. It's not just used to triggering OnDetached method, but also used to remove the pending runnables posted in the handlers(https://fburl.com/4spmzpwu).

Reviewed By: pasqualeanatriello

Differential Revision: D13848269

fbshipit-source-id: 0060ddb8cc0341cf7963e0b5fb4982bc1f9675f4