add 'doesNotHaveVisibleText' into litho view & component assertion'
Summary:
I was writing tests for a text component that the content is determined by conditions. Obviously, I need to test that the content not in the target branch isn's showed up unexpectedly. Then I found we don't have the counter function for `hasVisibleText` like we had for `hasVisibleDrawable`.

I understood that testing text not show up is harder than drawable or test key because sub-string is counted too in this test framework, but the demand is real and test key isn't applicable to `MountSpec`.

Anyway, we should provide the ability and flexibility in the assertion framework then left the test author to solve the potential problems.

After some code search, I found we do have `doesNotHaveVisibleText` for `ViewTree` assertion, which is used in `LIthoViewAssert`, which is used in `ComponentAssert` which is used in litho spec tests.

Reviewed By: pasqualeanatriello

Differential Revision: D14941223

fbshipit-source-id: 43322b50f03fb0b9f72000c3c247dcc9645bedc1