Make RenderInfoDebugInfoRegistry use weak entries
Summary:
According to https://docs.oracle.com/javase/7/docs/api/java/util/WeakHashMap.html, the keys of the weak hash map are weak, but the entries are not. Therefore if the entries retain a reference to context, the context will be retains, and therefore the views will be retained causing a memory leak.

This only happens on debug builds, so this doesn't affect prod. However, it does make it more difficult to do memory profiling on debug builds.

Reviewed By: astreet

Differential Revision: D14939401

fbshipit-source-id: 20332f815f74280deb3d7750fb8e381743b7d813