Add experiment param to kick off media layout creation ASAP
Summary:
We are seeing that we know what component to render for media during `initializeAdaptersAndControllers`, but we have to wait for measure to start computing those layouts. Additionally, since we don't want to compute the layout synchronously, we have to post a Runnable to the main thread to wait until after measure to do setRootAsync.

In this experiment, we use the assumption that the media will be full screen, and use that to kick off the async layout as soon as we know the media component to render.

Using systrace, this seems to move mounting of the video view (and thus starting of video playback) to earlier in TTRC.

Reviewed By: strulovich

Differential Revision: D14997285

fbshipit-source-id: 18afde5a424f4829289d3e4150d73bf5d4f22843