introduce DynamicValue
Summary:
We need an interface for developers to supply dynamic values to Litho components.

Requirements:

1. Should have a public setter method for setting/updating the value

2. Access to the `get()` method should be restricted - we need it internally, but don't want it be exposed to the users

3. We need to be able to subscribe/unsubscribe to the value changes. The dynamic value should accept multiple subscribers at the same time (one dynamic value could be passed to multiple components). We want to restrict access to subscription functionality, in order to minimize chances of people using the interface in the ways it is not supposed to be used

We introduce `DynamicValue<T>` class that meets all the requirements.
A `Component` object may act as a listener (will `OnValueChangeListener` interface) to multiple `DynamicValue`s. In order for `Component` to now what exactly `DynamicValues` has changed, we use cookie objects, supplied when we attach a listener, and passed back to the listener in callback method.

Reviewed By: astreet

Differential Revision: D14403872

fbshipit-source-id: 6af45db2235d5b9111252decf3d42e23d0944196