Resume tree creating in LayoutState
Summary:
When a LayoutState calculation is interrupted during the tree creation, we skip measuring and return the LayoutState with the partial InternalNode immediately.
When the UI thread will pick it up, it will pass in the partially calculated LayoutState, it will resume the InternalNode creation and will measure.

Not functional yet:
granularity around where exactly during LayoutState.calculate the interruption was triggered. Right now this is only able to resume from creating the InternalNode; I'll follow up with adding flags to be able to resume from InternalNode creation, measure or collectResults

Reviewed By: pasqualeanatriello

Differential Revision: D15149386

fbshipit-source-id: 6f98865b0a30c70876683dde30acd80d7309da89