Interrupt and resume LayoutState blocking UI
Summary:
When the UI thread needs the result of a LSF that's running in the BG thread, instead of blocking until that returns, the bg LSF is interrupted.
The LayoutState calculation will be finished on the UI thread.
Will follow up with a diff that can pick up the LayoutState calculation from anywhere it was interrupted, not just from the InternalNode creation

Reviewed By: pasqualeanatriello

Differential Revision: D15150785

fbshipit-source-id: da9eb30f543d0103aa217811ec6bc26e94e7f29a