Make all fields on LayoutOutput final
Summary:
The class LayoutOutput (https://fburl.com/v6jcxyyh) is meant to be an immutable object. The fields were not originally final because the object was originally pooled for efficiency on older versions of Android. Now that it isn't pooled, we should make the fields final since that lets the compiler and runtime do more optimizations!

This is more than just adding "final" on i-vars: it will mean doing a bit of refactoring in LayoutState.java to be able to set all the fields at once in the constructor.

Reviewed By: adityasharat

Differential Revision: D14952161

fbshipit-source-id: a310c41a6876547cf2487df63a76514fff387da8