Improve instrumentation tags
Summary:
Often we are getting tags where the summary is null and the source is just not meaningful enough to be actionable.
In this diff, I'm switching over to simple name and removing the null label to help human readers of the piped tags.

Before:
```
String tag = "calculateLayout 1 - null";
```

Now:
```
String tag = "calculateLayout " + <component-simple-name>;
```

Reviewed By: marco-cova

Differential Revision: D15366558

fbshipit-source-id: 21d78233862ca976c762515802d6a7680f5e4150