Remove mIsNestedTreeResolutionExperimentEnabled as a member variable from Component
Summary:
A `Component` can be created (`Component.Builder#create()`) with a `ComponentContext` which does not belong to the `ComponentTree`. This might import value of nested tree experiment config from another `ComponentTree` and break QE isolation.

In addition, the `if` condition in `makeShallowCopy` is no longer necessary because the key will be set again in `Component#updateInternalChildState()`.

This also us to remove the member variable and use the parent component context to get the correct value of the config.

Reviewed By: astreet

Differential Revision: D15448924

fbshipit-source-id: 5b45b903be2b0398c7ff850bbfbd3bee517ebc14