Only call OnCreateInitialState methods once
Summary:
this will guarantee that OnCreateInitialState can only be called once for any component global key inside a ComponentTree. The result will be commited immediately to the ComponentTree instead of waiting for a valid LayoutState to finish to commit the result.
Previously there were no guarantees around how many times it can be called if multiple threads are trying to init state at the same time.

Reviewed By: adityasharat

Differential Revision: D15416272

fbshipit-source-id: 9ae10ccc7a41628ee66ec94129da2f724bd3bc64