Rename InternalNode#getRootComponent to InternalNode#getTailComponent
Summary:
Rename `getRootComponent` to `getTailComponent`.

Consider the following component hierarchy:

```
                  A
                 / \
                E   B
                     \
                      C
                       \
                        D

```

When the layout is resolved the InternalNode tree will looks as follows

```
                 N1
                /  \
               N2   N3
```

The components and their global keys as follows:

```
N1 -> [
        A :  (A)
      ]

N2 -> [
        E :  (A,E)
      ]

N3 -> [
        0:  D -> "A,B,C,D",
        1:  C -> "A,B,C",
        2:  B -> "A,B",
      ]
```

We can see here that `D` which is at index `0` isn't really the root component of the InternalNode N3.

Reviewed By: muraziz

Differential Revision: D15469973

fbshipit-source-id: d26abf5a173d1899d8ec4636914d49d8ab7c1aa3